Finally, perhaps, reached a conclusion: in active inference integral control is necessary since there are unknown inputs by definition. Useless to compare the active inference implementation to standard LQG in this sense, since the approach is radically different and LQG fails miserably if inputs are not known. Better to embrace integral control and rather start asking how fundamental it must really be.
Funnily enough, the velocity form is ill defined if the integral action is not included (see Astrom 1995, pp 79-80) since, as I've seen, no unique stationary solution can be given (i.e. any solution will eventually give an error of zero).

So, stop trying to make active inference and LQG look alike, they will be different and for good reasons! Discuss the differences and include the idea that integral control seems to be everywhere if we assume no knowledge of input stimuli and a linear approximation of world dynamics.